+++ HISTORY.rst
@@ -149,6 +149,7 @@
 * Update ChoiceForm to be full height.
 * Add AddRemoveView to replace AssignRemoveView
 * Update the group roles view to use the new AddRemoveView.
+* Add role create and edit events.
 
 3.1.11 (2019-04-XX)
 ===================